STING/llm_service/requirements.gateway.txt
# Core API components
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
starlette>=0.27.0

# HTTP and networking
requests>=2.28.0
httpx>=0.24.0
aiohttp>=3.8.4

# Content filtering dependencies
detoxify>=0.5.1
regex>=2023.3.23
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0.0
python-multipart>=0.0.6

# Monitoring and observability
prometheus-client>=0.16.0
